fix tweet form for avatar and send button
